renamed to config-validation
