+++ frontend/src/scripts/config-validation.ts
